requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
ImageHash==4.3.1
iniconfig==2.0.0
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
scipy==1.13.1
six==1.16.0
